requirements.txt
# Guardian AI - Agent Orchestration System
# Complete requirements for fresh build
# Python 3.8+ required

# ============================================
# Core LangChain Framework
# ============================================
langchain>=0.1.0
langchain-core>=0.1.1
langchain-community>=0.0.20
langchain-google-genai>=0.0.11
langchain-text-splitters>=0.0.1

# ============================================
# LLM Provider - Google Gemini
# ============================================
google-generativeai>=0.3.2

# ============================================
# Vector Stores & Embeddings
# ============================================
chromadb>=0.4.22          # For legal document analysis
faiss-cpu>=1.7.4          # For code indexing and similarity search

# ============================================
# Document Processing
# ============================================
pypdf>=4.1.0              # PDF parsing for legal documents
pypdf2>=3.0.0             # Additional PDF support

# ============================================
# Git & Repository Tools
# ============================================
gitpython>=3.1.0          # Git repository cloning and analysis

# ============================================
# Web & HTTP
# ============================================
requests>=2.31.0          # HTTP requests
urllib3>=2.0.0            # URL handling

# ============================================
# Data Processing
# ============================================
numpy>=1.24.0             # Numerical computing
pandas>=2.0.0             # Data manipulation (optional but useful)

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0      # Environment variable management
pydantic>=2.0.0           # Data validation
tenacity>=8.2.0           # Retry logic
typing-extensions>=4.5.0  # Type hints support

# ============================================
# Database & Storage
# ============================================
sqlalchemy>=2.0.0         # Database ORM (for ChromaDB)
sqlite-vec>=0.0.1         # SQLite vector extension (optional)

# ============================================
# Async Support
# ============================================
aiohttp>=3.8.0            # Async HTTP client
asyncio>=3.4.3            # Async I/O support

# ============================================
# Optional: Alternative LLM Providers
# ============================================
# Uncomment if you want to use OpenAI or Anthropic
# langchain-openai>=0.0.5
# openai>=1.0.0
# langchain-anthropic>=0.1.0
# anthropic>=0.7.0
# tiktoken>=0.5.2

# ============================================
# Development & Testing (Optional)
# ============================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# black>=23.0.0            # Code formatter
# flake8>=6.0.0            # Linter
# mypy>=1.0.0              # Type checker
